+++ src/helpers.php
@@ -14,7 +14,7 @@ if (! function_exists('strings')) {
      * without a __toString method.
      *
      * @param mixed  $string   Value to modify, after being cast to string. Default: ''
-     * @param string $encoding The character encoding. Default: UTF-8
+     * @param mixed  $encoding The character encoding. Default: UTF-8
      */
     function strings($string = '', $encoding = 'UTF-8'): Strings
     {